backend/requirements.txt
# Django Core
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-filter==24.1

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter
dj-database-url==2.1.0  # Database URL parsing

# CORS for React Native
django-cors-headers==4.3.1

# Image handling
Pillow==10.2.0

# Environment variables
python-decouple==3.8

# Email
django-anymail==10.2  # For email services

# API Documentation
drf-yasg==1.21.7  # Swagger/OpenAPI

# Utilities
python-dateutil==2.8.2

# Production server
gunicorn==21.2.0

# Firebase Admin SDK (for push notifications)
firebase-admin==6.4.0
